Remove AbstractControlRequest as per review comment
